Move some code to a utils package.
This allows functions that REST endpoints call to be reused elsewhere.
